sprint-6/desafio/etapa-1/requirements.txt
boto3==1.34.134
python-dotenv==1.0.1
pandas==2.2.2